Fixing tests and refactoring
